suggest adding --index to stash snapshot
This PR suggests adding `--index` to the `git apply` command
that ends the `git snapshot` alias.

The reason is that otherwise, although, as you rightly say,
the working tree is restored so that it seems nothing has changed,
the index is destroyed.

But this way, there is no change in either the working tree
or the index, so it truly appears that the snapshot was taken
with _no_ visible effect of _any_ kind.